feat: moved router to a separate file as a class
